Move and slightly refactor common components
