+++ web/client/src/components/common/soro-button.vue
@@ -23,10 +23,3 @@ export default defineComponent({
     emits: ['click'],
 });
 </script>
-
-<style scoped>
-.soro-button {
-    margin-top: 0.2em;
-    margin-bottom: 0.2em;
-}
-</style>
