+++ web/client/src/components/common/soro-collapsible.vue
